pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ine-stat-scraper"
version = "0.1.0"
description = "Scraper para datos estadísticos de INE.Stat Chile - Módulos Agua y Aire"
authors = [
    {name = "Tu Nombre", email = "tu.email@example.com"},
]
dependencies = [
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "aiofiles>=23.0.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.43",
    "psycopg2-binary>=2.9.10",
]
requires-python = ">=3.9"

[tool.hatch.build.targets.wheel]
packages = ["src", "config"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
